requirements.txt
streamlit>=1.52.0
pandas>=2.0.0
requests>=2.31.0
lxml>=5.0.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
python-docx>=1.1.0
plotly>=5.18.0
APScheduler>=3.10.0
reportlab>=4.0.0
graphviz>=0.20.0
Pillow>=10.0.0
streamlit-agraph>=0.0.45
pyvis>=0.3.2
